requirements.txt
graia-ariadne
openai
pyyaml
loguru 